+++ sdk/azcore/internal/pollers/loc/loc.go
@@ -41,7 +41,7 @@ func CanResume(token map[string]interface{}) bool {
 
 // Poller is an LRO poller that uses the Location pattern.
 type Poller[T any] struct {
-	pl   exported.Pipeline
+	pl   exported.Policy
 	resp *http.Response
 
 	Type     string `json:"type"`
@@ -51,7 +51,7 @@ type Poller[T any] struct {
 
 // New creates a new Poller from the provided initial response.
 // Pass nil for response to create an empty Poller for rehydration.
-func New[T any](pl exported.Pipeline, resp *http.Response) (*Poller[T], error) {
+func New[T any](pl exported.Policy, resp *http.Response) (*Poller[T], error) {
 	if resp == nil {
 		log.Write(log.EventLRO, "Resuming Location poller.")
 		return &Poller[T]{pl: pl}, nil