+++ sdk/azcore/internal/pollers/util.go
@@ -269,7 +269,7 @@ func (p *NopPoller[T]) Result(ctx context.Context, out *T) error {
 // If the request fails, the update func is not called.
 // The update func returns the state of the operation for logging purposes or an error
 // if it fails to extract the required state from the response.
-func PollHelper(ctx context.Context, endpoint string, pl exported.Pipeline, update func(resp *http.Response) (string, error)) error {
+func PollHelper(ctx context.Context, endpoint string, pl exported.Policy, update func(resp *http.Response) (string, error)) error {
 	req, err := exported.NewRequest(ctx, http.MethodGet, endpoint)
 	if err != nil {
 		return err